refactor(invoice-pdf): extract shared PDF builder
Move buildInvoicePdf out of invoice-modal-table.tsx into a standalone
module so the "Save Invoice" flow, the multi-page test button, and
(next) exporting a previously saved invoice all render identically
instead of maintaining separate copies.
