+++ frontend/src/pages/admin/invoice/new-invoice/_components/invoice-modal-table.tsx
@@ -1,7 +1,10 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { toast } from "sonner";
-import jsPDF from "jspdf";
+import {
+  buildInvoicePdf,
+  InvoicePdfLineItem,
+} from "@/features/invoice/invoice-pdf";
 import {
   useInvoicePayloadQuery,
   useSelectedPayloadInvoiceQuery,
@@ -238,6 +241,7 @@ const FulfillmentTooltip = ({ invoiceItem, product }: TooltipProps) => {
 
 export const InvoiceTable = ({ invoiceNumber }: InvoiceTableProps) => {
   const navigate = useNavigate();
+  const [notes, setNotes] = useState("");
   const [discountValue, setDiscountValue] = useState("0");
   const [discountType, setDiscountType] = useState<"%" | "amount">("%");
   const [isSaving, setIsSaving] = useState(false);
@@ -262,298 +266,6 @@ export const InvoiceTable = ({ invoiceNumber }: InvoiceTableProps) => {
   const formatDiscount = (discount: number, isPercentage: boolean): string => {
     if (discount === 0) return "—";
     return isPercentage ? `${discount}%` : `₱${formatCurrency(discount)}`;
-  };
-
-  type InvoicePdfLineItem = {
-    label: string;
-    quantity: number;
-    unit: string;
-    unitPrice: number;
-    total: number;
-  };
-
-  type InvoicePdfOptions = {
-    invoiceNumberLabel: string;
-    customerName: string;
-    term: string;
-    dateLabel: string;
-    lineItems: InvoicePdfLineItem[];
-    subtotal: number;
-    discountAmount: number;
-    discountLabel: string;
-    grandTotal: number;
-    fileName: string;
-  };
-
-  // Shared PDF builder used both by the real "Save Invoice" flow and the
-  // dummy-data multi-page test flow, so both always render identically.
-  const buildInvoicePdf = ({
-    invoiceNumberLabel,
-    customerName,
-    term,
-    dateLabel,
-    lineItems,
-    subtotal,
-    discountAmount,
-    discountLabel,
-    grandTotal,
-    fileName,
-  }: InvoicePdfOptions) => {
-    const doc = new jsPDF({
-      orientation: "portrait",
-      unit: "mm",
-      format: "a4",
-    });
-    const pageWidth = doc.internal.pageSize.getWidth();
-    const pageHeight = doc.internal.pageSize.getHeight();
-    const left = 10;
-    const right = pageWidth - 10;
-    const contentWidth = right - left;
-    const bottomLimit = pageHeight - 10;
-
-    doc.setTextColor(0);
-    doc.setDrawColor(0);
-
-    const infoLabelX = right - 55;
-
-    // TABLE geometry (shared across pages)
-    const colWidths = { qty: 22, unit: 20, unitPrice: 26, total: 30 };
-    const descWidth =
-      contentWidth -
-      (colWidths.qty + colWidths.unit + colWidths.unitPrice + colWidths.total);
-    const colX = {
-      descStart: left,
-      qtyEnd: left + descWidth + colWidths.qty,
-      unitStart: left + descWidth + colWidths.qty + 2,
-      unitPriceEnd:
-        left + descWidth + colWidths.qty + colWidths.unit + colWidths.unitPrice,
-      totalEnd: right,
-    };
-    const lineHeight = 4.2;
-    const descTextWidth = descWidth - 3;
-
-    doc.setFont("helvetica", "normal");
-    doc.setFontSize(9);
-
-    // Pre-compute each row's wrapped description + height once, so both the
-    // dry-run (page counting) and real render pass use identical values.
-    // rowHeight is the height of the TEXT itself (line-count * lineHeight for
-    // a wrapped description) — it does NOT include any breathing room around
-    // it, that's handled separately by ROW_VERTICAL_PADDING below.
-    const rows = lineItems.map((item) => {
-      const descLines: string[] = doc.splitTextToSize(
-        item.label,
-        descTextWidth,
-      );
-      const rowHeight = Math.max(lineHeight, descLines.length * lineHeight);
-      return { ...item, descLines, rowHeight };
-    });
-
-    const FIRST_PAGE_TABLE_START_Y = 46.5; // after full header + table header
-    const CONTINUATION_TABLE_START_Y = 33.5; // after simplified header + table header
-    const CONTINUED_TEXT_RESERVE = 10; // room for "(Continued on Page N)"
-    const FOOTER_RESERVE = discountAmount > 0 ? 62 : 55; // grand total + signatures (+ discount lines)
-
-    // Total vertical space reserved around each row's text block (split
-    // above/below via centering) before the next divider line is drawn.
-    // This is deliberately generous so the divider never touches descenders
-    // (g, y, p, etc.) and single-line values look centered in the row.
-    const ROW_VERTICAL_PADDING = 6;
-    // Approx. distance from the vertical center of a text's line-box down
-    // to its baseline, for 9pt Helvetica — used to convert a "center Y" into
-    // the Y jsPDF actually expects (which is always a baseline).
-    const TEXT_BASELINE_CENTER_OFFSET = 1.3;
-
-    // ── PASS 1: dry run — figure out how many pages this invoice needs ──
-    let dryY = FIRST_PAGE_TABLE_START_Y;
-    let totalPages = 1;
-    rows.forEach((row, i) => {
-      const rowConsumption = row.rowHeight + ROW_VERTICAL_PADDING;
-      const isLastRow = i === rows.length - 1;
-      const reserve = isLastRow ? FOOTER_RESERVE : CONTINUED_TEXT_RESERVE;
-      if (dryY + rowConsumption + reserve > bottomLimit) {
-        totalPages += 1;
-        dryY = CONTINUATION_TABLE_START_Y;
-      }
-      dryY += rowConsumption;
-    });
-
-    // ── PASS 2: real render, now that totalPages is known ──
-    let y = 14;
-    let currentPage = 1;
-
-    const renderFirstPageHeader = () => {
-      y = 14;
-      const detailRow = (
-        leftLabel: string,
-        leftValue: string,
-        rightLabel: string,
-        rightValue: string,
-      ) => {
-        doc.setFont("helvetica", "bold");
-        doc.setFontSize(9.5);
-        doc.text(leftLabel, left, y);
-        doc.setFont("helvetica", "normal");
-        doc.text(leftValue || "-", left + 22, y);
-
-        doc.setFont("helvetica", "bold");
-        doc.text(rightLabel, infoLabelX, y);
-        doc.setFont("helvetica", "normal");
-        doc.text(rightValue || "-", right, y, { align: "right" });
-
-        y += 6.5;
-      };
-
-      detailRow("Customer:", customerName, "DR/Invoice #:", invoiceNumberLabel);
-      detailRow("Term:", term, "Date:", dateLabel);
-
-      doc.setFont("helvetica", "bold");
-      doc.setFontSize(9.5);
-      doc.text("Page:", infoLabelX, y);
-      doc.setFont("helvetica", "normal");
-      doc.text(`1 of ${totalPages}`, right, y, { align: "right" });
-
-      y += 4;
-      doc.setLineWidth(0.4);
-      doc.line(left, y, right, y);
-      y += 7;
-    };
-
-    const renderContinuationHeader = () => {
-      y = 14;
-      doc.setFont("helvetica", "bold");
-      doc.setFontSize(9.5);
-      doc.text(`DR/INV-${invoiceNumberLabel}`, left, y);
-      doc.text(`Page ${currentPage} of ${totalPages}`, right, y, {
-        align: "right",
-      });
-
-      y += 4;
-      doc.setLineWidth(0.4);
-      doc.line(left, y, right, y);
-      y += 7;
-    };
-
-    const renderTableHeader = () => {
-      doc.setFont("helvetica", "bold");
-      doc.setFontSize(9.5);
-      doc.text("Description", colX.descStart, y);
-      doc.text("Quantity", colX.qtyEnd, y, { align: "right" });
-      doc.text("Unit", colX.unitStart, y);
-      doc.text("Unit Price", colX.unitPriceEnd, y, { align: "right" });
-      doc.text("Total", colX.totalEnd, y, { align: "right" });
-      y += 3;
-      doc.setLineWidth(0.4);
-      doc.line(left, y, right, y);
-      y += 5.5;
-    };
-
-    const renderContinuedFooter = (nextPage: number) => {
-      doc.setFont("helvetica", "italic");
-      doc.setFontSize(8.5);
-      doc.text(`(Continued on Page ${nextPage})`, pageWidth / 2, bottomLimit, {
-        align: "center",
-      });
-    };
-
-    renderFirstPageHeader();
-    renderTableHeader();
-
-    rows.forEach((row, i) => {
-      const rowConsumption = row.rowHeight + ROW_VERTICAL_PADDING;
-      const isLastRow = i === rows.length - 1;
-      const reserve = isLastRow ? FOOTER_RESERVE : CONTINUED_TEXT_RESERVE;
-
-      if (y + rowConsumption + reserve > bottomLimit) {
-        renderContinuedFooter(currentPage + 1);
-        doc.addPage();
-        currentPage += 1;
-        renderContinuationHeader();
-        renderTableHeader();
-      }
-
-      doc.setFont("helvetica", "normal");
-      doc.setFontSize(9);
-
-      // The row occupies [y, rowBottom). Center every column's text
-      // vertically within that box: single-line columns center around the
-      // box's midline directly; the (possibly multi-line) description is
-      // centered as a whole block, so its first line starts higher up when
-      // it wraps to 2+ lines.
-      const rowBottom = y + rowConsumption;
-      const centerY = (y + rowBottom) / 2;
-      const singleLineY = centerY + TEXT_BASELINE_CENTER_OFFSET;
-      const descFirstLineY =
-        centerY - row.rowHeight / 2 + lineHeight / 2 + TEXT_BASELINE_CENTER_OFFSET;
-
-      doc.text(row.descLines, colX.descStart, descFirstLineY);
-      doc.text(String(row.quantity), colX.qtyEnd, singleLineY, {
-        align: "right",
-      });
-      doc.text(row.unit, colX.unitStart, singleLineY);
-      doc.text(formatCurrency(row.unitPrice), colX.unitPriceEnd, singleLineY, {
-        align: "right",
-      });
-      doc.text(formatCurrency(row.total), colX.totalEnd, singleLineY, {
-        align: "right",
-      });
-
-      doc.setLineWidth(0.2);
-      doc.line(left, rowBottom, right, rowBottom);
-      y = rowBottom;
-    });
-
-    // ── FOOTER (Subtotal/Discount + Grand Total + Signatures) — LAST PAGE ONLY ──
-    y += 5;
-
-    if (discountAmount > 0) {
-      doc.setFont("helvetica", "normal");
-      doc.setFontSize(10);
-      doc.text("Subtotal:", right - 70, y);
-      doc.text(`P${formatCurrency(subtotal)}`, right, y, { align: "right" });
-      y += 5.5;
-
-      doc.text("Discount:", right - 70, y);
-      doc.text(`-${discountLabel}`, right, y, { align: "right" });
-      y += 6.5;
-
-      doc.setLineWidth(0.3);
-      doc.line(right - 70, y - 4.5, right, y - 4.5);
-    }
-
-    doc.setFont("helvetica", "bold");
-    doc.setFontSize(12);
-    doc.text("GRAND TOTAL:", right - 70, y);
-    doc.text(`P${formatCurrency(grandTotal)}`, right, y, { align: "right" });
-
-    y += 20;
-
-    // SIGNATURE BLOCKS
-    const signatureWidth = contentWidth / 3 - 4;
-    const signatureLineHeight = 3.5;
-    const signatureBlock = (x: number, label: string) => {
-      doc.setFont("helvetica", "bold");
-      doc.setFontSize(8);
-      const labelLines: string[] = doc.splitTextToSize(label, signatureWidth);
-      doc.text(labelLines, x, y);
-
-      const lineY = y + labelLines.length * signatureLineHeight + 6;
-      doc.setLineWidth(0.3);
-      doc.line(x, lineY, x + signatureWidth, lineY);
-
-      doc.setFont("helvetica", "normal");
-      doc.setFontSize(7);
-      doc.text("Signature Above Printed Name", x, lineY + 4);
-    };
-
-    signatureBlock(left, "\nPrepared By:");
-    signatureBlock(left + signatureWidth + 6, "\nChecked By:");
-    signatureBlock(
-      left + (signatureWidth + 6) * 2,
-      "Received the Above Goods in Good Order and Condition:",
-    );
-
-    doc.save(fileName);
   };
 
   const generateInvoicePdf = () => {
@@ -637,6 +349,7 @@ export const InvoiceTable = ({ invoiceNumber }: InvoiceTableProps) => {
         selectedCustomer?.id,
         user?.user_ID,
         invoiceTerm,
+        notes,
       );
       if (!res) {
         toast.dismiss(toastId);
@@ -804,6 +517,17 @@ export const InvoiceTable = ({ invoiceNumber }: InvoiceTableProps) => {
             No items added
           </div>
         )}
+      </div>
+
+      <div className="flex flex-col gap-1">
+        <label className="text-sm font-semibold">Notes</label>
+        <textarea
+          className="bg-custom-gray rounded-lg px-4 py-3 text-sm resize-none"
+          rows={3}
+          placeholder="Add any notes for this invoice..."
+          value={notes}
+          onChange={(e) => setNotes(e.target.value)}
+        />
       </div>
 
       <div className="flex justify-between items-end pt-2">
